refactor : Dynamic Routing에 따른 routing 변경
